separation mail sending responsibilty to a util file and utilizing return of client.send() method of mailtrap api
